pyproject.toml
[project]
name = "mona-cli"
version = "0.1.0"
description = "CLI tool for syncing local config files to a Lisa remote config server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[project.scripts]
mona = "mona.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mona"]

[tool.basedpyright]
typeCheckingMode = "standard" # or "strict", "basic"
# Example: Override specific diagnostic rules
reportUnusedImport = "warning"
# Example: Define execution environments for complex layouts
extraPaths = ["."]

